+++ packages/database/src/index.ts
@@ -2,3 +2,4 @@ export { getPrismaClient, type PrismaClient } from "./client";
 export { createPrismaIdentityRepositories } from "./repositories";
 export { createPrismaPropertyRepositories } from "./property-repositories";
 export { createPrismaAnalysisRepository } from "./analysis-repositories";
+export { createPrismaReviewTransaction } from "./review-transaction";